+++ Assets/Prefab/FinalScenePrefab/GreenArmor.prefab
@@ -92,8 +92,8 @@ BoxCollider:
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 3, y: 3, z: 3}
-  m_Center: {x: 0, y: 0, z: 0}
+  m_Size: {x: 1.0613599, y: 1.1991584, z: 1.2897587}
+  m_Center: {x: -0.011753321, y: 0.021064162, z: -0.05335331}
 --- !u!114 &3161642301143769346
 MonoBehaviour:
   m_ObjectHideFlags: 0
