+++ Assets/Prefab/FinalScenePrefab/Healthpack.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &478646108806776149
+--- !u!1 &6801511175590676314
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,49 +8,49 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 478646108806776170}
-  - component: {fileID: 478646108806776171}
-  - component: {fileID: 478646108806776168}
-  - component: {fileID: 478646108806776169}
-  - component: {fileID: 478646108806776150}
-  - component: {fileID: 478646108806776151}
-  - component: {fileID: 478646108806776148}
+  - component: {fileID: 6801511175590676304}
+  - component: {fileID: 6801511175590676319}
+  - component: {fileID: 6801511175590676318}
+  - component: {fileID: 6801511175590676317}
+  - component: {fileID: 6801511175590676316}
+  - component: {fileID: 6801511175590676315}
+  - component: {fileID: 6801511175590676305}
   m_Layer: 0
-  m_Name: BlueArmor
+  m_Name: Healthpack
   m_TagString: HealthPack
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &478646108806776170
+--- !u!4 &6801511175590676304
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 3.98, y: 0.74, z: -2.83}
+  m_LocalPosition: {x: 1.79, y: 0.74, z: -3.81}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &478646108806776171
+--- !u!33 &6801511175590676319
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &478646108806776168
+--- !u!23 &6801511175590676318
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -81,38 +81,38 @@ MeshRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
---- !u!65 &478646108806776169
+--- !u!65 &6801511175590676317
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_Material: {fileID: 0}
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 3, y: 3, z: 3}
-  m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &478646108806776150
+  m_Size: {x: 1.4714165, y: 1.2853549, z: 1.6914811}
+  m_Center: {x: -0.032219887, y: -0.0066660643, z: 0.014240503}
+--- !u!114 &6801511175590676316
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: ae4fc4799c70e45cba9bcccaa1be582f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!54 &478646108806776151
+--- !u!54 &6801511175590676315
 Rigidbody:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   serializedVersion: 2
   m_Mass: 1
   m_Drag: 0
@@ -122,17 +122,17 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &478646108806776148
+--- !u!114 &6801511175590676305
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 478646108806776149}
+  m_GameObject: {fileID: 6801511175590676314}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 3843f98f64aa7409d87ce99657bfdde4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   thisGun: {fileID: 0}
-  thisHealth: {fileID: 11400000, guid: dd28bbba472e34b579a1564506dccc7a, type: 2}
+  thisHealth: {fileID: 11400000, guid: b117857d39fe0479f84b933ead64e225, type: 2}
