+++ Assets/Prefab/FinalScenePrefab/Pistol_Weapon.prefab.meta
