+++ Assets/Prefab/UnusedPrefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 971dee1a4e13b46598f6b064eed89652
-NativeFormatImporter:
+guid: 7bbf2ce2397b37941aac63da1c4c20a9
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
