--- Assets/ScriptableObjects/Ammunition.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ccaa67a49b124006aae172258a6f4ad6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
