+++ Assets/Scripts/GunHealthManager.cs
@@ -4,18 +4,7 @@ using UnityEngine;
 
 public class GunHealthManager : MonoBehaviour
 {
-    public Guns thisGun;
+    public Guns thisGun; 
     public HealthObjects thisHealth;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
 }
