+++ ProjectSettings/TimeManager.asset
@@ -3,7 +3,7 @@
 --- !u!5 &1
 TimeManager:
   m_ObjectHideFlags: 0
-  Fixed Timestep: 0.008333334
+  Fixed Timestep: 0.016666668
   Maximum Allowed Timestep: 1
   m_TimeScale: 1
   Maximum Particle Timestep: 1
